+++ Echo/src/utopia/echo/model/request/OllamaRequest.scala
@@ -58,7 +58,10 @@ trait OllamaRequest[+R] extends ApiRequest[R]
 			"model" -> llm.llmName,
 			"stream" -> stream,
 			"options" -> optionsModel.notEmpty)
-		Left((baseModel ++ customProperties).withoutEmptyValues)
+		val bodyModel = (baseModel ++ customProperties).withoutEmptyValues
+		// TODO: Remove test prints
+		println(s"Outgoing request body: $bodyModel")
+		Left(bodyModel)
 	}